Various minor changes
