+++ HandlingEditor.Client/HandlingEditor.Client.csproj
@@ -53,7 +53,7 @@
     <Compile Include="HandlingEditor.cs" />
     <Compile Include="HandlingPreset.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="Utilities.cs" />
+    <Compile Include="Enumerables.cs" />
     <Compile Include="VehiclesPermissions.cs" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
